Refactor browser test
